requirements.txt
fastapi[all]
uvicorn
python-dotenv
pydantic[email]
bcrypt==4.0.1
pyjwt==2.8.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0
redis==5.0.1
motor==3.3.2
beanie==1.25.0
pymongo==4.6.0
minio==7.2.7
python-multipart==0.0.9